backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
google-api-core>=2.20.0
yfinance==0.2.32
requests==2.31.0
